requirements.txt
streamlit
pandas
SQLAlchemy
python-dotenv
haystack-ai
haystack-google-genai